+++ packages/@dopt/javascript-common/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @dopt/javascript-common
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [f636dd8b]
+  - @dopt/block-types@1.0.2
+
+## 1.2.0
+
+### Minor Changes
+
+- 850b6fe7: update optimisticUpdates parameter, clarifying what it does and where to use it
+
 ## 1.1.0
 
 ### Minor Changes